requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
gspread==5.10.0
google-auth==2.23.0
validators==0.22.0
gspread